+++ zerver/decorator.py
@@ -4,22 +4,9 @@ import logging
 import urllib
 from functools import wraps
 from io import BytesIO
-from typing import (
-    TYPE_CHECKING,
-    Callable,
-    Dict,
-    Optional,
-    Sequence,
-    Set,
-    TypeVar,
-    Union,
-    cast,
-    overload,
-)
+from typing import TYPE_CHECKING, Callable, Dict, Optional, Sequence, TypeVar, Union, cast, overload
 
 import django_otp
-import orjson
-from circuitbreaker import CircuitBreakerError, circuit
 from django.conf import settings
 from django.contrib.auth import REDIRECT_FIELD_NAME
 from django.contrib.auth import login as django_login
@@ -38,7 +25,6 @@ from django_otp import user_has_device
 from two_factor.utils import default_device
 from typing_extensions import Concatenate, ParamSpec
 
-from zerver.lib.cache import cache_with_key
 from zerver.lib.exceptions import (
     AccessDeniedError,
     AnomalousWebhookPayload,
@@ -50,7 +36,6 @@ from zerver.lib.exceptions import (
     OrganizationAdministratorRequired,
     OrganizationMemberRequired,
     OrganizationOwnerRequired,
-    RateLimited,
     RealmDeactivatedError,
     RemoteServerDeactivatedError,
     UnauthorizedError,
@@ -59,7 +44,7 @@ from zerver.lib.exceptions import (
     WebhookError,
 )
 from zerver.lib.queue import queue_json_publish
-from zerver.lib.rate_limiter import RateLimitedIPAddr, RateLimitedUser
+from zerver.lib.rate_limiter import is_local_addr, rate_limit, rate_limit_user
 from zerver.lib.request import REQ, RequestNotes, has_request_variables
 from zerver.lib.response import json_method_not_allowed, json_success
 from zerver.lib.subdomains import get_subdomain, user_matches_subdomain
@@ -70,16 +55,10 @@ from zerver.lib.utils import has_api_key_format, statsd
 from zerver.models import Realm, UserProfile, get_client, get_user_profile_by_api_key
 
 if settings.ZILENCER_ENABLED:
-    from zilencer.models import (
-        RateLimitedRemoteZulipServer,
-        RemoteZulipServer,
-        get_remote_server_by_uuid,
-    )
+    from zilencer.models import RemoteZulipServer, get_remote_server_by_uuid
 
 if TYPE_CHECKING:
     from django.http.request import _ImmutableQueryDict
-
-rate_limiter_logger = logging.getLogger("zerver.lib.rate_limiter")
 
 webhook_logger = logging.getLogger("zulip.zerver.webhooks")
 webhook_unsupported_events_logger = logging.getLogger("zulip.zerver.webhooks.unsupported")
@@ -907,10 +886,6 @@ def authenticated_json_view(
     return _wrapped_view_func
 
 
-def is_local_addr(addr: str) -> bool:
-    return addr in ("127.0.0.1", "::1")
-
-
 # These views are used by the main Django server to notify the Tornado server
 # of events.  We protect them from the outside world by checking a shared
 # secret, and also the originating IP (for now).
@@ -918,16 +893,6 @@ def is_local_addr(addr: str) -> bool:
 def authenticate_notify(request: HttpRequest, secret: str = REQ("secret")) -> bool:
     return is_local_addr(request.META["REMOTE_ADDR"]) and constant_time_compare(
         secret, settings.SHARED_SECRET
-    )
-
-
-def client_is_exempt_from_rate_limiting(request: HttpRequest) -> bool:
-
-    # Don't rate limit requests from Django that come from our own servers,
-    # and don't rate-limit dev instances
-    client = RequestNotes.get_notes(request).client
-    return (client is not None and client.name.lower() == "internal") and (
-        is_local_addr(request.META["REMOTE_ADDR"]) or settings.DEBUG_RATE_LIMITING
     )
 
 
@@ -990,100 +955,6 @@ def statsd_increment(
         return wrapped_func
 
     return wrapper
-
-
-def rate_limit_user(request: HttpRequest, user: UserProfile, domain: str) -> None:
-    """Returns whether or not a user was rate limited. Will raise a RateLimited exception
-    if the user has been rate limited, otherwise returns and modifies request to contain
-    the rate limit information"""
-
-    RateLimitedUser(user, domain=domain).rate_limit_request(request)
-
-
-@cache_with_key(lambda: "tor_ip_addresses:", timeout=60 * 60)
-@circuit(failure_threshold=2, recovery_timeout=60 * 10)
-def get_tor_ips() -> Set[str]:
-    if not settings.RATE_LIMIT_TOR_TOGETHER:
-        return set()
-
-    # Cron job in /etc/cron.d/fetch-tor-exit-nodes fetches this
-    # hourly; we cache it in memcached to prevent going to disk on
-    # every unauth'd request.  In case of failures to read, we
-    # circuit-break so 2 failures cause a 10-minute backoff.
-
-    with open(settings.TOR_EXIT_NODE_FILE_PATH, "rb") as f:
-        exit_node_list = orjson.loads(f.read())
-
-    # This should always be non-empty; if it's empty, assume something
-    # went wrong with writing and treat it as a non-existent file.
-    # Circuit-breaking will ensure that we back off on re-reading the
-    # file.
-    if len(exit_node_list) == 0:
-        raise OSError("File is empty")
-
-    return set(exit_node_list)
-
-
-def rate_limit_ip(request: HttpRequest, ip_addr: str, domain: str) -> None:
-    RateLimitedIPAddr(ip_addr, domain=domain).rate_limit_request(request)
-
-
-def rate_limit_request_by_ip(request: HttpRequest, domain: str) -> None:
-    # REMOTE_ADDR is set by SetRemoteAddrFromRealIpHeader in conjunction
-    # with the nginx configuration to guarantee this to be *the* correct
-    # IP address to use - without worrying we'll grab the IP of a proxy.
-    ip_addr = request.META["REMOTE_ADDR"]
-    assert ip_addr
-
-    try:
-        # We lump all TOR exit nodes into one bucket; this prevents
-        # abuse from TOR, while still allowing some access to these
-        # endpoints for legitimate users.  Checking for local
-        # addresses is a shortcut somewhat for ease of testing without
-        # mocking the TOR endpoint in every test.
-        if is_local_addr(ip_addr):
-            pass
-        elif ip_addr in get_tor_ips():
-            ip_addr = "tor-exit-node"
-    except (OSError, CircuitBreakerError) as err:
-        # In the event that we can't get an updated list of TOR exit
-        # nodes, assume the IP is _not_ one, and leave it unchanged.
-        # We log a warning so that this endpoint being taken out of
-        # service doesn't silently remove this functionality.
-        rate_limiter_logger.warning("Failed to fetch TOR exit node list: %s", err)
-        pass
-    rate_limit_ip(request, ip_addr, domain=domain)
-
-
-def rate_limit_remote_server(
-    request: HttpRequest, remote_server: "RemoteZulipServer", domain: str
-) -> None:
-    try:
-        RateLimitedRemoteZulipServer(remote_server, domain=domain).rate_limit_request(request)
-    except RateLimited as e:
-        rate_limiter_logger.warning(
-            "Remote server %s exceeded rate limits on domain %s", remote_server, domain
-        )
-        raise e
-
-
-def rate_limit(request: HttpRequest) -> None:
-    if not settings.RATE_LIMITING:
-        return
-
-    if client_is_exempt_from_rate_limiting(request):
-        return
-
-    user = request.user
-    remote_server = RequestNotes.get_notes(request).remote_server
-
-    if settings.ZILENCER_ENABLED and remote_server is not None:
-        rate_limit_remote_server(request, remote_server, domain="api_by_remote_server")
-    elif not user.is_authenticated:
-        rate_limit_request_by_ip(request, domain="api_by_ip")
-    else:
-        assert isinstance(user, UserProfile)
-        rate_limit_user(request, user, domain="api_by_user")
 
 
 def return_success_on_head_request(
